Set Exercise.weight as double, truncate display to 2 decimals

+++ lib/models/workout.dart
@@ -12,10 +12,10 @@ class Workout {
 
 class Exercise {
   final String name;
-  final int weight;
+  final double weight;
 
   Exercise({
    required this.name,
    required this.weight,
-});
+  });
 }
